requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
pandas
python-dotenv
scikit-learn

